requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
SQLAlchemy==2.0.36
alembic==1.14.0
requests==2.32.3
pymongo==4.10.1

python-dotenv==1.0.1

# Test
pytest==8.3.4